Add script that automatically generates 'latest' dirs for each plugin
Adds script `setup_latest_dirs.sh` that can be used to automatically
latest meta.yamls in each plugin dir. The added meta.yamls are copies of
the meta.yaml with newest 'firstPublishedDate', with version set to be
'latest'.

If a latest directory already exists and a newer version is present
for that plugin, a warning is printed.

Only applies to the v3 registry path.

Also adds script to dockerfiles to ensure that each plugin has a
'latest' version.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>