Add latest version for existing plugins
Adds 'latest' version for every existing plugin, which is a copy of the
current latest version with version field overwritten to be 'latest'.

For che-theia, since plugin version 1.0.0 actually points at the
'latest' dockerimage, it is instead renamed to 'latest'. For the
remaining plugins, the newest version listed is unchanged.

Signed-off-by: Angel Misevski <amisevsk@redhat.com>