Add script to automatically detect and file issues for out of date plugins
This PR contributes a Python script that:
	-reads all plugins in a plugin registry
	-checks their versions
	-checks the upstream repositories of their vsix files
	-checks whether the plugin's vsix file needs to be updated
	-prints out a list of plugins that need to be updated

A Dockerfile is also attached to this PR, and provides an environment
to run/edit the script.

Additionally, this script can also file GitHub issues for any plugin
that needs to be updated. A README with installation/usage instructions
has been provided.

Fixes eclipse/che##16215

Signed-off-by: Eric Williams <ericwill@redhat.com>