Validate meta.yaml versions to be strings, not numbers (#427)
* Validate meta.yaml versions to be strings, not numbers

Do not allow plugin versions to be specifid in a way that parses as a
number. Supporting number alongside string can complicate
deserialization implementations, as it's not obvious when a value could
be parsed as a number. Examples:
- 3.2.1 is a string, but 3.2 requires quotes to parse as a string
- Git commit hashes typically contain letters (parse as string) but
occasionally will be exclusively digits (parse as number).

Signed-off-by: Angel Misevski <amisevsk@redhat.com>

* Fix how plugin IDs are computed during the build

Fix an issue where computing plugin ids fails if they are strings
enclosed in quotation marks

Signed-off-by: Angel Misevski <amisevsk@redhat.com>